Add Basic Boyer-Moore-Horspool Implementaion
A basic boyer-moore-horspool function has been added to the main.cpp
it uses this algorithm to search through a text and print out the
positions the pattern is found. This implementaion is just the basic
building block of this algorithm, so needs so improvement.

As part of the implementation the CharTable class was added to allow
the pre-processing of the pattern string and calcutate the required
skip steps needed for the algorithm to function.