Basic Implementation Of Sequencial And Parallel Search
A very basic implementation of a sequencial search of a list of
patterns in the now very long test text and a parallel implementation
this is just assigning a thread to each pattern and searching through
the text.

The parallel search is showing an improvement even with this basic
implementation, it seems to have more effect as the test text gets
longer.