Add Task Based Busy Waiting (Farm & Worker) Search
To attempt to improve the performance a busy waiting apprach has
been implemented. This works by having a TaskFarm of Tasks that are
executed over a set number of threads. The TaskFarm runs these using
WorkerThreads that are continually kept busy until the task_queue in
the farm has been processed.

An abstract Task class has also been added so in theory the farm can
process any inherited Task Classes, this might be useful later if
other tasks are needed.

The SearchTask uses the Boyer Moore Horspool algorithm to search for
a pattern in a text.