Add StringSearcher Class
The StringSearcher Class contains all of the search methods to
abstract them from the main(). This also prevents the need for the
pattern list and search text to be copied multiple times.

All the searches in main() have been replaced with the StringSearcher
equivalent. The SearchTask has also been modified to add a RunParallel()
to allow the capture of the results from this SearchTask when executed
on separate threads.

All the issue #1 opened as there is no method of specifying the
number of threads used in StringSearch::SearchParallelSimple().