Refactor Main() and Add GetPatternList() to StringSearcher
All of the code in the main() has been refactored so all the major
elements are in separate funtions to make the code more readable in
preparation to the automated timing code is added.

To encapsulate these functions a little more the GetPatternList()
was added to the StringSearcher so the pointer for the pattern
list is is retrieved from the StringSearcher to prevent errors.