Resolve Memory Leak In TaskFarm (Fixes #3)
The TaskFarm has a memory leak where tasks were created on the heap
but then the vector storing the pointers was cleared before the
threads where cleared from the heap. Instead the clean up was
attempted on the un-initialized vector causing not onlly a memory
leak but runtime issues. Fixes #3.