Results From Searcher Stored In Unordered_Map (Fixes #2)
To ensure that the results form the task based parallel search are
correcly matched with their associated pattern string the results
are now stored in an unordered_map to maintain this.

The ConsoleUI was also updated to allow the printing of this new
unordered_map results.

Longer real world search texts added and a pattern list of the top
100 most common English words.